pmdm697-yard/pm97django/version-history/djangosite697-v02/requirements.txt
Django>=4
django-extensions
djangorestframework
django-htmx

djangorestframework_simplejwt
django-cors-headers
#django-crontab
apscheduler
gunicorn

mysqlclient
django-environ
whitenoise
psycopg2
